[IEE-72] Remove deprecated api calls
- Improve naming scheme of check interval

+++ core/domain/src/main/kotlin/com/kastik/apps/core/domain/service/WorkScheduler.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 
 interface WorkScheduler {
-    fun scheduleAnnouncementAlerts(intervalMinutes: Long)
+    fun scheduleAnnouncementAlerts(intervalMinutes: Int)
     fun getAnnouncementAlertsWorkInfo(): Flow<WorkInfo?>
     fun cancelAnnouncementAlerts()
 }
